:truck: Refactor SeriesApiUnitTest, SeriesRemoteDataSourceUnitTest and SeriesRepositoryUnitTest directories
